move InitializeGame and ContextEnhancer into separate files
