+++ packages/core.js
@@ -7,11 +7,8 @@
  */
 
 import { Game } from '../src/core/game.js';
-import {
-  InitializeGame,
-  CreateGameReducer,
-  INVALID_MOVE,
-} from '../src/core/reducer.js';
+import { InitializeGame } from '../src/core/initialize.js';
+import { CreateGameReducer, INVALID_MOVE } from '../src/core/reducer.js';
 import { Flow, FlowWithPhases } from '../src/core/flow.js';
 import { Pass, TurnOrder } from '../src/core/turn-order.js';
 import { PlayerView } from '../src/core/player-view.js';
